app/requirements.txt
Flask-SQLAlchemy>=3.1
Flask-Login>=0.6
Flask-Migrate>=4.0
Flask-WTF>=1.2
email-validator>=2.0
